moved cli prompt into internal folder
