+++ src/goose/_internal/cli/prompt/completer.py
