+++ src/goose/_internal/cli/prompt/lexer.py
