+++ src/goose/_internal/cli/prompt/prompt_validator.py
