+++ src/goose/_internal/cli/prompt/user_input.py
